Added saved encryption and easy encryption
You can now use a shortcut (Ctrl + Shift + E) to encrypt with the
password that you use being stored for each domain. Right click on the
text to decrypt it. For decryption you still need to enter a password
everytime. Also the context menu can allow you to clear or set password
for the domain.

I need to add a feature that allows you to change whether you want to
save passwords and allow an option to save passwords when decrypting.

Known issues:
- Editing a decrypt text and editing that will cause the yellow
  background to keep
- Still doesn't work on sites like Discord (Clicking off will reset the
  text)